+++ src/main/java/com/checkmarx/flow/controller/bitbucket/server/BitbucketServerController.java
@@ -21,6 +21,10 @@ import com.checkmarx.flow.dto.bitbucketserver.PullEvent;
 import com.checkmarx.flow.dto.bitbucketserver.PushEvent;
 import com.checkmarx.flow.exception.InvalidTokenException;
 import com.checkmarx.flow.exception.MachinaRuntimeException;
+import com.checkmarx.flow.handlers.bitbucket.server.BitbucketServerEventHandler;
+import com.checkmarx.flow.handlers.bitbucket.server.BitbucketServerMergeHandler;
+import com.checkmarx.flow.handlers.bitbucket.server.BitbucketServerPushHandler;
+import com.checkmarx.flow.handlers.config.ConfigContextProvider;
 import com.checkmarx.flow.service.BitBucketService;
 import com.checkmarx.flow.service.ConfigurationOverrider;
 import com.checkmarx.flow.service.CxScannerService;
