+++ src/main/java/com/checkmarx/flow/handlers/bitbucket/server/BitbucketServerMergeHandler.java
@@ -1,5 +1,4 @@
-package com.checkmarx.flow.controller.bitbucket.server;
-
+package com.checkmarx.flow.handlers.bitbucket.server;
 import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
@@ -17,19 +16,26 @@ import lombok.NonNull;
 import lombok.experimental.SuperBuilder;
 
 @SuperBuilder
-public class BitbucketServerPushHandler extends BitbucketServerEventHandler {
+public class BitbucketServerMergeHandler extends BitbucketServerEventHandler {
 
-    private static final Logger log = org.slf4j.LoggerFactory.getLogger(BitbucketServerPushHandler.class);
-
-    @NonNull
-    protected String branchFromRef;
+    private static final Logger log = org.slf4j.LoggerFactory.getLogger(BitbucketServerMergeHandler.class);
 
     @NonNull
-    protected String toHash;
+    private String currentBranch;
+
+    @NonNull
+    private String targetBranch;
+    
+    @NonNull
+    private String fromRefLatestCommit;
+
+    @NonNull
+    private String pullRequestId;
 
 
     @Override
     public ResponseEntity<EventResponse> execute(String uid) {
+
         controllerRequest = webhookUtils.ensureNotNull(controllerRequest);
 
         try {
@@ -37,29 +43,37 @@ public class BitbucketServerPushHandler extends BitbucketServerEventHandler {
                 application = controllerRequest.getApplication();
             }
 
-            // set the default bug tracker as per yml
-            webhookUtils.setBugTracker(configProvider.getFlowProperties(), controllerRequest);
-            BugTracker.Type bugType = ScanUtils.getBugTypeEnum(controllerRequest.getBug(),
-                    configProvider.getFlowProperties().getBugTrackerImpl());
-
-            Optional.ofNullable(controllerRequest.getAppOnly())
-                    .ifPresent(configProvider.getFlowProperties()::setTrackApplicationOnly);
+            BugTracker.Type bugType = BugTracker.Type.BITBUCKETSERVERPULL;
+            if (!ScanUtils.empty(controllerRequest.getBug())) {
+                bugType = ScanUtils.getBugTypeEnum(controllerRequest.getBug(), configProvider.getFlowProperties().getBugTrackerImpl());
+            }
+            Optional.ofNullable(controllerRequest.getAppOnly()).ifPresent(configProvider.getFlowProperties()::setTrackApplicationOnly);
 
             if (ScanUtils.empty(product)) {
                 product = ScanRequest.Product.CX.getProduct();
             }
             ScanRequest.Product p = ScanRequest.Product.valueOf(product.toUpperCase(Locale.ROOT));
-            String currentBranch = ScanUtils.getBranchFromRef(branchFromRef);
             List<String> branches = webhookUtils.getBranches(controllerRequest, configProvider.getFlowProperties());
-            String latestCommit = toHash;
 
-            BugTracker bt = ScanUtils.getBugTracker(controllerRequest.getAssignee(), bugType,
-                    configProvider.getJiraProperties(), controllerRequest.getBug());
-            FilterConfiguration filter = configProvider.getFilterFactory().getFilter(controllerRequest,
-                    configProvider.getFlowProperties());
+            BugTracker bt = ScanUtils.getBugTracker(controllerRequest.getAssignee(), bugType, 
+              configProvider.getJiraProperties(), controllerRequest.getBug());
+
+            FilterConfiguration filter = configProvider.getFilterFactory().getFilter(controllerRequest, 
+              configProvider.getFlowProperties());
 
             String gitUrl = getGitUrl();
             String gitAuthUrl = getGitAuthUrl(gitUrl);
+
+            String repoSelfUrl = getRepoSelfUrl(toProjectKey, toSlug);
+
+            String mergeEndpoint = repoSelfUrl.concat(MERGE_COMMENT);
+            mergeEndpoint = mergeEndpoint.replace("{id}", pullRequestId);
+
+            String buildStatusEndpoint = configProvider.getBitBucketProperties().getUrl().concat(BUILD_API_PATH);
+            buildStatusEndpoint = buildStatusEndpoint.replace("{commit}", fromRefLatestCommit);
+
+            String blockerCommentUrl = repoSelfUrl.concat(BLOCKER_COMMENT);
+            blockerCommentUrl = blockerCommentUrl.replace("{id}", pullRequestId);
 
             ScanRequest request = ScanRequest.builder().application(application).product(p)
                     .project(controllerRequest.getProject())
@@ -70,21 +84,27 @@ public class BitbucketServerPushHandler extends BitbucketServerEventHandler {
                     .repoUrlWithAuth(gitAuthUrl)
                     .repoType(ScanRequest.Repository.BITBUCKETSERVER)
                     .branch(currentBranch)
+                    .mergeTargetBranch(targetBranch)
+                    .mergeNoteUri(mergeEndpoint)
                     .refs(refId)
-                    .email(emails)
-                    .scanPreset(controllerRequest.getPreset())
+                    .email(null)
                     .incremental(controllerRequest.getIncremental())
+                    .scanPreset(controllerRequest.getPreset())
                     .excludeFolders(controllerRequest.getExcludeFolders())
                     .excludeFiles(controllerRequest.getExcludeFiles())
                     .bugTracker(bt)
                     .filter(filter)
-                    .hash(latestCommit)
+                    .hash(fromRefLatestCommit)
                     .build();
 
             setBrowseUrl(request);
             fillRequestWithCommonAdditionalData(request, toProjectKey, toSlug, webhookPayload);
             checkForConfigAsCode(request);
+            request.putAdditionalMetadata("buildStatusUrl", buildStatusEndpoint);
+            request.putAdditionalMetadata("cxBaseUrl", configProvider.getCxScannerService().getProperties().getBaseUrl());
+            request.putAdditionalMetadata("blocker-comment-url", blockerCommentUrl);
             request.setId(uid);
+
             // only initiate scan/automation if target branch is applicable
             if (configProvider.getHelperService().isBranch2Scan(request, branches)) {
                 configProvider.getFlowService().initiateAutomation(request);
@@ -94,4 +114,5 @@ public class BitbucketServerPushHandler extends BitbucketServerEventHandler {
         }
         return webhookUtils.getSuccessMessage();
     }
+
 }
